requirements.txt
numpy== 1.24.0
scipy==1.8.0
matplotlib==3.5.1
pillow==9.0.1
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
scikit_learn==1.5.0
pandas==2.2.2
tqdm==4.66.2

#matplotlib==3.8.4
#numpy== 2.1.1
#scipy==1.13.0
#pillow==10.3.0